Added endpoint to view list of new connections for Admin - Committed by Shreya Nair

+++ src/main/java/com/connection/emobile/dto/UserRequestDto.java
@@ -2,18 +2,18 @@ package com.connection.emobile.dto;
 
 public class UserRequestDto {
 
-	private int mobileNumber;
+	private Long mobileNumber;
 	private String username;
 	private String address;
 	private String emailId;
 	private String panCardNumber;
 	private int planId;
 
-	public int getMobileNumber() {
+	public Long getMobileNumber() {
 		return mobileNumber;
 	}
 
-	public void setMobileNumber(int mobileNumber) {
+	public void setMobileNumber(Long mobileNumber) {
 		this.mobileNumber = mobileNumber;
 	}
