+++ src/main/java/com/connection/emobile/repository/MobileNumberRepository.java
@@ -2,10 +2,11 @@ package com.connection.emobile.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import com.connection.emobile.entity.MobileNumber;
 
 @Repository
-public interface MobileNumberRepository extends JpaRepository<MobileNumber, Integer> {
+public interface MobileNumberRepository extends JpaRepository<MobileNumber, Long> {
+	
+	 
 
 }
